+++ pytsite/assetman/_error.py
@@ -1,0 +1,14 @@
+"""PytSite Assetman Error
+"""
+
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+class PackageNotRegistered(Exception):
+    pass
+
+
+class PackageAlreadyRegistered(Exception):
+    pass